+++ frontend/src/components/TvShowsBubble.js
@@ -11,7 +11,7 @@ const TvShowsBubble = ({ count, res, data }) => {
   const goToDetail = () => {
     if (Object.keys(res).length > 0) {
       router("/my-search-detail", {
-        state: { res: res, data: data, storyType: "TV Show" },
+        state: { res: res, data: data, storyType: "TV Shows" },
       });
     }
   };