+++ frontend/src/pages/MarketPlace.js
@@ -1,6 +1,11 @@
 import React from "react";
+
+//******* Import component(s) *******/
 import FooterComponent from "../components/FooterComponent";
 import Navbar from "../components/NavBar";
+
+//******* Import style logic *******/
+//Here
 
 export default function MarketPlace() {
   return (