+++ frontend/src/pages/MyAccount.js
@@ -1,7 +1,11 @@
 import React from "react";
-// Connecting the NavBar to the pages
+
+//******* Import component(s) *******/
 import FooterComponent from "../components/FooterComponent";
 import Navbar from "../components/NavBar";
+
+//******* Import style logic *******/
+//Here
 
 export default function MyAccount() {
   return (